.internal/pycli/requirements.txt
ruamel.yaml ~= 0.16.12
blessed ~= 1.17.5
pyyaml ~= 5.3.1
requests ~= 2.22.0
asyncio ~= 3.4.3
